stage1: Add compiler-rt cross-compilation
This rather complicated step is only run when we're cross-compiling
architectures. We first have to provide some system-target headers
through a mini-bootstrap of musl, and use the host clang and config-
musl to build the compiler-rt builtins.

These builtins are required to actually link musl in the next step
of stage1, which we'll eventually use to rebootstrap the toolchain
in stage2.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>