stage2.5: Utilise new mount helpers in base.sh
These ensure we have /dev/pts, /sys and /proc available for use within
our chroot environment, which experience has shown are absolutely essential
for most software to build and run correctly.

We will also need to flesh out the missing device nodes in /dev with `mknod`
but that will be a separate change.

Our intent is to NEVER bindmount /dev/ as this can lead to catastrophic
data loss if used incorrectly.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>