stage3/coreutils: Do not define Serpent OS as 'GNU/Linux'
Serpent OS is built using a multitude of technologies and
origins, and the core toolchain is musl + LLVM projects.
While we do ship some GNU software (usually as dependencies
of another item of GNU software) we're not defined as
a GNU/Linux distribution, and are binary incompatible.

Let's dispell any misconceptions early of compatibility
that does not and will not exist.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>