stage3/systemd: Partially unblock building
We have to disable anything using gshadow, but that still
fails to build, as stuff still tries to use it when disabled..

We have an upstreamed patch for includes, and set a define
to ensure if_ether.h doesn't completely foobar the build.

Other bits are undesirable yet or requiring more build
dependencies.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>